Refactor the client code: move all App logic from index.jsx into App.jsx
