+++ client/src/components/carousel.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+
+const Carousel = ({ similarListings }) => (
+  <div>
+    <h2>Similar Listings</h2>
+    <h2>Test: {similarListings[0].photo}</h2>
+  </div>
+);
+
+export default Carousel;
